fix(persistence): use i32 for remaining Postgres int4 bulk-submit columns
The first int4/i64 fix covered get_manifest and process_entries, but the
async worker path had the same mismatch on other INTEGER columns:

- bulk_entry_results.line_number (INTEGER): read as i64 in
  get_entry_results and bound as i64 in the entry-result INSERT.
- bulk_manifests.processed_entries / failed_entries (INTEGER): bound as
  i64 in update_manifest_progress.

tokio-postgres rejects the int4<->i64 conversion. update_manifest_progress
runs after the worker ingests each file, so the bind error aborted run_job
and left the manifest stuck `processing`; it was then re-leased and failed
again indefinitely, so the submission never reached `complete` and
$bulk-submit-status returned 202 forever (smoke poll loop exhausted, exit 1).
SQLite is dynamically typed and unaffected.

Read/bind these three INTEGER columns as i32. BIGINT columns
(fencing_token, last_processed_line, line_count, byte_count) and
COUNT/SUM aggregates remain i64.

Tests: cargo test -p helios-persistence --lib bulk_submit (28 passed); fmt
and CI-style clippy clean. The postgres round trip itself is covered only
by the external smoke workflow (no local Docker).

+++ crates/persistence/src/backends/postgres/bulk_submit.rs
@@ -779,7 +779,7 @@ impl BulkSubmitProvider for PostgresBackend {
         let results: Vec<BulkEntryResult> = rows
             .iter()
             .map(|row| {
-                let line_number: i64 = row.get(0);
+                let line_number: i32 = row.get(0);
                 let resource_type: String = row.get(1);
                 let resource_id: Option<String> = row.get(2);
                 let created: Option<bool> = row.get(3);
@@ -973,7 +973,7 @@ impl PostgresBackend {
                     &submission_id.submitter.as_str(),
                     &submission_id.submission_id.as_str(),
                     &manifest_id,
-                    &(result.line_number as i64),
+                    &(result.line_number as i32),
                     &result.resource_type.as_str(),
                     &result.resource_id,
                     &result.created,
@@ -1477,8 +1477,8 @@ impl SubmitWorkerStorage for PostgresBackend {
                  WHERE tenant_id = $4 AND submitter = $5 AND submission_id = $6
                    AND manifest_id = $7 AND worker_id = $8 AND fencing_token = $9",
                 &[
-                    &(processed_entries as i64),
-                    &(failed_entries as i64),
+                    &(processed_entries as i32),
+                    &(failed_entries as i32),
                     &(last_processed_line as i64),
                     &lease.tenant.tenant_id().as_str(),
                     &lease.submission_id.submitter,
